pyproject.toml
[project]
name = "interier-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.128.0",
    "google>=3.0.0",
    "google-genai>=1.61.0",
    "jinja2>=3.1.6",
    "langchain>=1.2.8",
    "langchain-openai>=1.1.7",
    "openai>=2.16.0",
    "pillow>=12.1.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
